+++ src/Phan/Language/UnionType.php
@@ -48,7 +48,7 @@ class UnionType implements \Serializable
     // phpcs:ignore Generic.NamingConventions.UpperCaseConstantName
     const union_type_regex =
         Type::type_regex
-        . '(\|' . Type::type_regex . ')*';
+        . '(\s*\|\s*' . Type::type_regex . ')*';
 
     /**
      * @var string
@@ -61,7 +61,7 @@ class UnionType implements \Serializable
     // phpcs:ignore Generic.NamingConventions.UpperCaseConstantName
     const union_type_regex_or_this =
         Type::type_regex_or_this
-        . '(\|' . Type::type_regex_or_this . ')*';
+        . '(\s*\|\s*' . Type::type_regex_or_this . ')*';
 
     /**
      * @var array<int,Type> * This is an immutable list of unique types.